implement BlifParser
