+++ src/Compiler/CMakeLists.txt
@@ -80,6 +80,7 @@ set (SRC_CPP_LIST
   QLDeviceManager.cpp
   QLSettingsManager.cpp
   QLMetricsManager.cpp
+  BlifParser.cpp
 )
 
 set (SRC_H_INSTALL_LIST
@@ -111,6 +112,7 @@ set (SRC_H_INSTALL_LIST
   QLDeviceManager.h
   QLSettingsManager.h
   QLMetricsManager.h
+  BlifParser.h
 )
 
 set (SRC_H_LIST
